+++ packages/breadboard/src/inspector/node.ts
@@ -1,0 +1,43 @@
+/**
+ * @license
+ * Copyright 2024 Google LLC
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+import { NodeDescriptor } from "../types.js";
+import { InspectableGraph, InspectableNode } from "./types.js";
+
+export const inspectableNode = (
+  descriptor: NodeDescriptor,
+  inspectableGraph: InspectableGraph
+): InspectableNode => {
+  return new Node(descriptor, inspectableGraph);
+};
+
+class Node implements InspectableNode {
+  descriptor: NodeDescriptor;
+  #graph: InspectableGraph;
+  #incoming: InspectableNode[] | undefined;
+  #outgoing: InspectableNode[] | undefined;
+
+  constructor(descriptor: NodeDescriptor, graph: InspectableGraph) {
+    this.descriptor = descriptor;
+    this.#graph = graph;
+  }
+
+  incoming(): InspectableNode[] {
+    return (this.#incoming ??= this.#graph.incomingForNode(this.descriptor.id));
+  }
+
+  outgoing(): InspectableNode[] {
+    return (this.#outgoing ??= this.#graph.outgoingForNode(this.descriptor.id));
+  }
+
+  isEntry(): boolean {
+    return this.incoming().length === 0;
+  }
+
+  isExit(): boolean {
+    return this.outgoing().length === 0;
+  }
+}
